% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipTab.R
\name{clipTab}
\alias{clipTab}
\title{Olito's function para copiar una tabla al portapapeles}
\usage{
clipTab(df)
}
\arguments{
\item{df}{data.frame o data.table apegar}
}
\value{
character de aviso
}
\description{
Pega una data.frame o data.table al portapapeles para pegar facil al Excel
}
\examples{
# data(mtcars)
# clipTab(mtcars)

}
