% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatNum.R
\name{formatNum}
\alias{formatNum}
\title{Olito's function para formatear un número}
\usage{
formatNum(num, decimal = 0, sep = ".")
}
\arguments{
\item{num}{Número a formatear}

\item{decimal}{define el n de decimales a incluir (aproxima)}

\item{sep}{define el separador de miles, por defecto es punto.}
}
\value{
char con el resultado
}
\description{
Formatea numero intercalando comas en los miles
}
\examples{
# formatNum(123456789.287)
# formatNum("123456789.287")
# formatNum("123456789.987", decimal = 1)
# formatNum("123456789.987", decimal = 1, sep = ',')
}
