% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrut.R
\name{nrut}
\alias{nrut}
\title{Olito's function para calcular un distinct en una tabla}
\usage{
nrut(df, var, format = TRUE)
}
\arguments{
\item{df}{data.frame o data.table}

\item{var}{string variable a testear, debe ir entre comillas}

\item{format}{logical default TRUE para agregar separador de miles}
}
\value{
numero
}
\description{
Calcula el numero de elementos diferentes en una variable de un data.frame
}
\examples{
# data <- data.frame(id = rep(1:1234, 2), val = runif(1234*2))
# nrut(data, 'id')
# nrut(data, 'id', format = FALSE)

}
