% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oclip.R
\name{oclip}
\alias{oclip}
\title{Olito's function para copiar una objeto al portapapeles}
\usage{
oclip(obj_to_copy = NULL, tipo = "vector")
}
\arguments{
\item{obj_to_copy}{data.frame o vector a pegar}

\item{tipo}{tipo de objeto data.frame o vector a pegar}
}
\value{
character de aviso
}
\description{
Pega una data.frame o vector al portapapeles
}
\examples{
# data(mtcars)
# clipTab(mtcars, 'tab')

}
