% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opwcorr.r
\name{ocorr}
\alias{ocorr}
\title{Olito's Correlation}
\usage{
ocorr(
  target = NULL,
  indep = NULL,
  grpvar = NULL,
  data = NULL,
  type = "pearson"
)
}
\arguments{
\item{target}{target variables}

\item{indep}{independent variables, if NULL: target vars will fill this parameter}

\item{grpvar}{Group variable (only one)}

\item{data}{data frame}

\item{type}{Type of correltation "pearson" o "spearman"}
}
\value{
data frame
}
\description{
Calculate correlations by transforming variables into a matrix and then
use of hmisc package the rest is data management. Calculates pearson and spearman. Wrapper from Stata.
}
\examples{
# library(dplyr); library(Hmisc)
# ocorr("mpg", "hp", data = mtcars, type = "spearman")
# ocorr(c("mpg", "cyl", "qsec"), c("drat", "hp", "wt"), data = mtcars, grpvar = "am")
# ocorr(c("mpg", "cyl", "qsec"), data = mtcars)
}
