% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odir.R
\name{odir}
\alias{odir}
\title{Olito's function para ver el directorio filtrado}
\usage{
odir(ext = NULL, find = NULL)
}
\arguments{
\item{ext}{Extension con punto o sin punto, mayuscula o minuscula}

\item{find}{Texto a buscar archvio, da igual minusc o mayusc}
}
\value{
data.table con la columna filtrada
}
\description{
filtra del dir por extension o texto
}
\examples{
# odir(ext = 'xlsx')
# odir(find = 'oli')
# odir(ext = 'xlsx', find = 'oli')



}
