% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okeep.r
\name{okeep}
\alias{okeep}
\title{keep variables funcion}
\usage{
okeep(keep = NULL, dframe = FALSE, fun = FALSE, char = FALSE, num = FALSE)
}
\arguments{
\item{keep}{Character vector with the name of the objects to keep}

\item{dframe}{Logical, set TRUE for keep all "data frames" in the environment}

\item{fun}{Logical, set TRUE for keep all "functions" in the environment}

\item{char}{Logical, set TRUE for keep all "character objects" in the environment}

\item{num}{Logical, set TRUE for keep all "numeric objects" in the environment}
}
\value{
Just drop any object not included in the parameters
}
\description{
Esta funcion borra o conserva objetos del gloval environment se puede conservar objetos puntuales por nombre y por clase el resto se borra del entorno.
}
\examples{
# Create some objects and keep one and data frames
# cars <- mtcars
# a <- 1
# b <- 2
# c <- "qwerty"
# d <- data.frame(d = 123)
# e <- "asdfg"
# f <- TRUE
# okeep(keep = c("a", "f"), dframe = TRUE)

}
