% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordervar.r
\name{ordervar}
\alias{ordervar}
\title{Olito's OrderVar para ordenar variables}
\usage{
ordervar(datos, varmove, after = 1)
}
\arguments{
\item{datos}{Un data.frame}

\item{varmove}{o variables a mover (en un vector char)}

\item{after}{Después de cual variable se dese mover (char or index), default = 1}
}
\value{
Data frame con las variables ordenadas
}
\description{
Ordena variables dentro de un data.frame
}
\examples{
# Get some data and save order
# data(mtcars)
# nameberfore <- names(mtcars)
# Ordenar
# mtcars <- ordervar(mtcars, "mpg", "am")
# nameberfore; names(mtcars)
# mtcars <- ordervar(mtcars, c("gear", "carb"), after = "wt")

}
