% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osumm.r
\name{osumm}
\alias{osumm}
\title{Olitos's Summary, descriptives}
\usage{
osumm(numvar, grpvar = NULL, data, rnd = 2, clip = FALSE)
}
\arguments{
\item{numvar}{Variable o Vector de variables en String}

\item{grpvar}{Variable de agrupacion, puede ser un string, factor o numerico, si no esta no se usa}

\item{data}{Data frame}

\item{rnd}{Numero de decimales}

\item{clip}{Clipboard option}
}
\value{
Devuelve un data frame que ademas se pasa al portapepeles
}
\description{
Esta funcion calcula estadisticas descriptivas de una variable numerica y en forma grupal, tiene version corta y extensa y de clip al porta papeles.
}
\examples{
# numvar <- "mpg"
# numvar <- c("mpg", "disp", "wt")
# grpvar <- "am"
# grpvar <- c("am", "vs")
# osumm(numvar, data = mtcars)
# osumm(numvar, grpvar, data = mtcars)
}
