% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otable.r
\name{otable}
\alias{otable}
\title{Calcula una tabla de contingencia entre 2 variables}
\usage{
otable(rvar = NULL, cvar = NULL, data = NULL, clip = 0)
}
\arguments{
\item{rvar}{Variable que ir? en las filas de la tabla}

\item{cvar}{Variable que ir? en las columnas de la tabla}

\item{data}{Data Frame que contiene las variables}

\item{clip}{Opciones de portapapeles, 0 = return, 1 = freq, 2 = row, 3 = col, 4 = col, 5 = chi2, 6 = 1 & 5}
}
\value{
Devuelve una dataframe si 1 var o lista con las 4 tablas y el test si contingencia. Seg?n clip se pasa al clipboard.
}
\description{
Funcion que calcula una tabla de contingencia entre dos variables de cualquier tipo las cuales se pasan a factor. Calcula la frecuencia, porcentaje por fila y columna y celda, ademas de test de independencia Chi-cuadrado y prueba exacta de Fisher.
}
\examples{
# otable(rvar = "cyl", data = mtcars)
# otable(rvar = "cyl", cvar = "am", data = mtcars, clip = 0)
# otable(rvar = "cyl", cvar = "am", data = mtcars, clip = 1)

}
