% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otext_clean.R
\name{otext_clean}
\alias{otext_clean}
\title{Olito Text Cleaner}
\usage{
otext_clean(txt)
}
\arguments{
\item{txt}{texto a limpiar}
}
\value{
char texto limpio
}
\description{
Quita las puntuaciones, espacios, tildes, numeros de un un texto.
}
\examples{
otext_clean("texto con tildes")
}
