% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ottest.indep.R
\name{ottest.indep}
\alias{ottest.indep}
\title{Olito's ttest para variables independientes.}
\usage{
ottest.indep(
  varlist = NULL,
  grpvar = NULL,
  data = NULL,
  version = 1,
  save = "file"
)
}
\arguments{
\item{varlist}{Vector String de variables dependientes}

\item{grpvar}{Variable (string) de grupo}

\item{data}{Data frame con los datos}

\item{version}{tipo de resultado}

\item{save}{Guarda un excel en el wd}
}
\value{
Devuelve un data frame con los resultados
}
\description{
Prueba ttest directo a un dataframe para muestras independientes.
}
\examples{
# grpvar <- "am"
# depvar <- "hp"
# varlist <- c("mpg", "disp", "drat", "hp")
# ottest.indep(varlist, grpvar, mtcars)
}
