% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectWD.R
\name{selectWD}
\alias{selectWD}
\title{Olito's function para seleccionar wd}
\usage{
selectWD()
}
\value{
character con el path
}
\description{
Selecciona un working directory mediante un archivo, ejecuta el setwd() y guarda un objeto mainDir.
}
\examples{
# selectWD()
}
