% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simText.R
\name{simText}
\alias{simText}
\title{Olito's function buscar textos similar}
\usage{
simText(test = NULL, txtvector = NULL, excluye = c(NA, "", " "))
}
\arguments{
\item{test}{character, txto a buscar}

\item{txtvector}{character vector, es el vector en donde se van a buscar los similares}

\item{excluye}{vector con la excepciones para el texto, devuelve un NA}
}
\value{
character del valor más parecido
}
\description{
Dado un texto o palabra, busca el más similar en un vector de palabras. Excepto si el input es NA o "", usa la libreria "stringdist" con default en "osa". No la he probado vectorizada... solo digo!!!
}
\examples{
# Crear un vector de testeo con la data de "islands"
# txtvector <- names(datasets::islands)
# simText("4Fr1caaa", txtvector)
# simText("#", txtvector, excluye = "#")
}
