% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlMerge.R
\name{sqlMerge}
\alias{sqlMerge}
\title{Olito's SQLite Merge for large files}
\usage{
sqlMerge(x = NULL, y = NULL, by = NULL, drop = TRUE, replace = FALSE)
}
\arguments{
\item{x}{Xdf data.frame class}

\item{y}{Ydf data.frame class}

\item{by}{Key variable, same type and name accross Xdf and Ydf}

\item{drop}{Logical for remove original Xdf and Ydf}

\item{replace}{Logical ovewrite an existing db file}
}
\value{
data.table object
}
\description{
Merge implementation in SQLite for dealing with memory issues when normal merge functions get out of RAM
}
\examples{
# library(RSQLite)
# library(DBI)
# library(glue)
# library(data.table)
# temp <- mtcars
# temp$id <- row.names(temp)
# row.names(temp) <- NULL
# yDF <- temp[1:26, c(12, 1:5)]
# xDF <- temp[7:32, c(12, 6:11)]
# sqlMerge(xDF, yDF, by = 'id', replace = TRUE)
}
